Get read-level data for a region from a modBAM file.

This function returns a RangedSummarizedExperiment object with read-level footprinting data extracted for a specific region from a bamfile in modBAM format. getReadDataForRegion() is a convenience wrapper around three other functions that could also be called independently from one another:

modkitExtract

runs modkit and extracts the read-level base modification information from a modBAM file.

readModkitExtract

reads the resulting output from modkit into a SummarizedExperiment object.

filterReadData

calculates quality measures and optionally filters out data from low quality reads.

getReadDataForRegion(
  bamfile,
  region,
  arglist.modkitExtract = list(modkit_bin = "modkit"),
  arglist.readModkitExtract = list(),
  arglist.filterReadData = NULL,
  assay.type = "mod_prob",
  verbose = FALSE
)

Arguments

bamfile

Character scalar with the name of a modBAM file (a BAM file that stores modified-base data).

region

A GRanges object with a single region for which data should be extracted from bamfile. Alternatively, the region can be specified as a character scalar (e.g. "chr1:1200-1300") that can be coerced into a GRanges object.

arglist.modkitExtract

A named list with further argument to pass to modkitExtract (must not contain bamfile or regions, which are set automatically).

arglist.readModkitExtract

A named list with further argument to pass to readModkitExtract (must not contain fnames, which is set automatically).

arglist.filterReadData

A named list with further argument to pass to filterReadData. Alternatively, it can be set to NULL, in which case no filtering will be performed.

assay.type

A string scalar specifying the assay name to store modification probabilities.

verbose

If TRUE, report on progress.

Value

A RangedSummarizedExperiment object with positions of modified bases in rows and reads in the columns.

See also

Author

Michael Stadler

Examples

if (FALSE) { # \dontrun{
modbamfile <- system.file("extdata", "6mA_1_10reads.bam", package = "footprintR")
se <- getReadDataForRegion(modbamfile,
                           region = "chr1:6940000-6955000",
                           arglist.readModkitExtract = list(modbase = "a"))
se
} # }