Quantify windows by calculating the difference between modification fractions for the positive and negative strand

Calculate sums of per-position Nmod and Nvalid counts for positions that overlap windows in gr, in a strand-specific manner, and calculate FracMod as the ratio between the two. Then calculate the difference between the FracMod for the + and - strand.

Usage

strandDiffFracMod(se, gr, pseudocount = 0)

Arguments

se

A RangedSummarizedExperiment object, as generated by readModBam(..., level = "summary") (i.e. with the assays "Nmod" and "Nvalid").

gr

A GRanges object defining the windows to quantify.

pseudocount

A numeric scalar with a pseudocount that will be added to the "Nmod" and "Nvalid" before calculating the modification fraction.

Value

A RangedSummarizedExperiment object with one row per window defined by gr and the window quantification contained in the assay "FracModDiff".

Author

Charlotte Soneson